+++ scripts/scene-packer.js
@@ -4761,17 +4761,24 @@ export default class ScenePacker {
         switch (type) {
           case 'JournalEntryPacks':
             if (CONSTANTS.IsV10orNewer()) {
-              const pages = document.pages?.filter(p => p.type === 'text' && p.text?.content) || [];
+              const pages = document.pages?.filter(p => p.type === 'text' && (p.text?.content || p.text?.markdown)) || [];
               if (pages.length) {
                 const pageUpdates = [];
                 for (const page of pages) {
+                  let path = 'text.content';
+                  let originalContent = page.text.content;
+                  if (page.text.format === CONST.JOURNAL_ENTRY_PAGE_FORMATS.MARKDOWN) {
+                    path = 'text.markdown';
+                    originalContent = page.text.markdown;
+                  }
+
                   let {
                     newContent,
                     references,
                     hyperlinksChanged,
-                  } = await this.relinkContent(page.text.content, entry, page, {domParser, rex, moduleName, packs, pack, typeName});
+                  } = await this.relinkContent(originalContent, entry, page, {domParser, rex, moduleName, packs, pack, typeName});
 
-                  if (newContent !== page.text.content) {
+                  if (newContent !== originalContent) {
                     ScenePacker.logType(
                       moduleName,
                       'info',
@@ -4786,7 +4793,7 @@ export default class ScenePacker {
                     );
                     pageUpdates.push({
                       _id: page._id,
-                      'text.content': newContent,
+                      [path]: newContent,
                     })
                   }
                 }